backend/requirements.txt
flask
flask-cors
openai
python-dotenv
urllib3
gradio